requirements.txt
Django
psycopg2
Unipath
South
Pillow
python-mimeparse
django-tastypie
